# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Telegram Drive API",
    "description": "HTTP API for Telegram-backed object storage. Headless (Docker) exposes the full surface on one port (1334): legacy /upload, /metrics, WebDAV, static Web UI, and /d/* share downloads. Desktop optional REST API (127.0.0.1:8550) is a subset — see docs/DESKTOP-API.md. Authenticate with X-API-Key or X-Access-Pwd (Web admin / desktop local access password).",
    "version": "3.0.0-beta"
  },
  "servers": [{ "url": "/", "description": "Current host" }],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "AccessPwdAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Access-Pwd",
        "description": "Web admin console password (ACCESS_PWD)"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/transport": {
      "get": {
        "tags": ["System"],
        "summary": "Current Telegram transport mode (bot or user)",
        "security": [{ "AccessPwdAuth": [] }, { "ApiKeyAuth": [] }],
        "responses": { "200": { "description": "Transport info JSON" } }
      }
    },
    "/api/v1/transport/mode": {
      "post": {
        "tags": ["System"],
        "summary": "Switch transport mode at runtime",
        "security": [{ "AccessPwdAuth": [] }, { "ApiKeyAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "mode": { "type": "string", "enum": ["bot", "user"] } }
              }
            }
          }
        },
        "responses": { "200": { "description": "Mode updated" } }
      }
    },
...
    "/api/v1/auth/phone/request": {
      "post": {
        "tags": ["Auth"],
        "summary": "Request login code",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["phone"],
                "properties": { "phone": { "type": "string", "example": "+8613800138000" } }
              }
            }
          }
        },
        "responses": { "200": { "description": "Code sent" } }
      }
    },
    "/api/v1/auth/phone/sign-in": {
      "post": {
        "tags": ["Auth"],
        "summary": "Sign in with SMS code",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["code"],
                "properties": { "code": { "type": "string" } }
              }
            }
          }
        },
        "responses": { "200": { "description": "Logged in" } }
      }
    },
    "/api/v1/auth/phone/password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Sign in with 2FA password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["password"],
                "properties": { "password": { "type": "string" } }
              }
            }
          }
        },
        "responses": { "200": { "description": "Logged in" } }
      }
    },
...
  }
}